+++ src/modules/calendar/components/index.ts
@@ -1,0 +1,1 @@
+export * from './monthCalendar/monthCalendar.component';