+++ src/modules/calendar/directives/calendarDayTemplate/calendarDayTemplate.context.ts
@@ -1,0 +1,12 @@
+import {CalendarDayData} from '../../interfaces';
+
+/**
+ * Context passed to template of calendar day
+ */
+export interface CalendarDayTemplateContext<TDate = unknown, TEvent = unknown>
+{
+    /**
+     * Data that are used for displaying calendar day
+     */
+    $implicit: CalendarDayData<TDate, TEvent>;
+}