+++ src/modules/calendar/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './calendarDayTemplate/calendarDayTemplate.context';
+export * from './calendarDayTemplate/calendarDayTemplate.directive';