+++ src/modules/calendar/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './calendarDayData.interface';
+export * from './calendarEventDayMetadata.interface';
+export * from './eventData.interface';