+++ src/modules/calendar/misc/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './calendarDayAspectRatio.enum';
+export * from './monthCalendarDayFormat.enum';