+++ src/modules/calendar/misc/enums/monthCalendarDayFormat.enum.ts
@@ -1,0 +1,20 @@
+/**
+ * Available day formats for calendar day
+ */
+export enum MonthCalendarDayFormat
+{
+    /**
+     * No week day name displayed
+     */
+    None = 'None',
+
+    /**
+     * Short version of week day name
+     */
+    Short = 'Short',
+
+    /**
+     * Full version of week day name
+     */
+    Full = 'Full',
+}