+++ src/modules/index.ts
@@ -1,3 +1,4 @@
+export * from './calendar';
 export * from './datePipes.module';
 export * from './dateTime';
 export * from './dateTimePicker';